+++ client/src/components/Header/Header.module.scss
@@ -2,11 +2,11 @@
   display: flex;
   justify-content: space-between;
   // align-items: center;
-  padding: 15px 30px;
-  width: 100vw;
+  padding: 25px 30px;
+  width: 100%;
   position: fixed;
   top: 0;
-  z-index: 100;
+  z-index: 2;
 
   .clock {
     position: relative;
@@ -31,7 +31,7 @@
 
       .navText {
         color: black;
-        font-size: 26px;
+        font-size: 28px;
         position: relative;
 
         .pill {
@@ -41,11 +41,10 @@
           border: 1px solid black;
 
           .blockerL {
-            z-index: 0;
             position: absolute;
             top: 13%;
-            left: -2px;
-            width: 3px;
+            left: -4px;
+            width: 6px;
             height: 75%;
             background: #eceae9;
           }
@@ -53,27 +52,27 @@
           .blockerR {
             position: absolute;
             top: 13%;
-            right: -2px;
-            width: 3px;
+            right: -4px;
+            width: 6px;
             height: 75%;
             background: #eceae9;
           }
 
           .blockerT {
             position: absolute;
-            top: -2px;
+            top: -4px;
             left: 6%;
             width: 89%;
-            height: 3px;
+            height: 5px;
             background: #eceae9;
           }
 
           .blockerB {
             position: absolute;
-            bottom: -2px;
+            bottom: -4px;
             left: 6%;
             width: 89%;
-            height: 3px;
+            height: 6px;
             background: #eceae9;
           }
         }
