+++ client/src/components/Hero/FoldText/FoldText.module.scss
@@ -1,15 +1,13 @@
 .textContainer {
-  width: 100vw;
+  width: 100%;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 25px 30px;
+  padding: 12px 30px;
 
 
   .statusTextContainer {
     font-size: 20px;
-    position: relative;
-    top: 0px;
     width: 333px;
