+++ client/src/components/Hero/Hero.module.scss
@@ -9,13 +9,11 @@
       z-index: 1;
       mix-blend-mode: multiply;
       position: absolute;
-      left: 0;
-      right: 0;
-      top: 0;
-      bottom: 0;
-      margin: auto;
+      top: 2%;
+      left: 50%;
+      transform: translateX(-50%);
       width: 50vw;
-      height: 80vh;
+      height: 90vh;
       max-width: 1200px;
     }
 
@@ -32,7 +30,7 @@
 
     .foldTextContainer {
       position: absolute;
-
+      width: 100%;
       bottom: 0;
     }
   }
