+++ client/src/components/Hero/TextScroll/TextScroll.jsx
@@ -1,6 +1,6 @@
 import classes from "./TextScroll.module.scss";
 import { useProgress } from "@react-three/drei";
-import { kimchi } from "@/components/Fonts/Fonts";
+import { kimchi } from "@/components/Fonts";
 
 export default function TextScroll() {
   const { progress } = useProgress();
