+++ client/src/components/Layout/TextReveal/TextReveal.module.scss
@@ -1,0 +1,5 @@
+.text {
+  font-size: 46px;
+  color: white;
+
+}
