+++ client/src/components/WipHero/FoldText/FoldText.module.scss
@@ -8,8 +8,6 @@
 
   .statusTextContainer {
     font-size: 20px;
-    position: relative;
-    top: 0px;
     width: 333px;
