+++ client/src/pages/contact/index.jsx
@@ -1,6 +1,7 @@
 import styles from "@/styles/Home.module.scss";
 import Head from "next/head";
-import WipHero from "@/components/WipHero/WipHero";
+import WipHero from "@/components/WipHero";
+import Layout from "@/components/Layout";
 
 export default function Home() {
   return (
@@ -11,9 +12,11 @@ export default function Home() {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main className={styles.main}>
-        <WipHero />
-      </main>
+      <Layout>
+        <main className={styles.main}>
+          <WipHero />
+        </main>
+      </Layout>
     </>
   );
 }
