+++ client/src/pages/index.jsx
@@ -1,7 +1,7 @@
 import styles from "@/styles/Home.module.scss";
 import Head from "next/head";
-import Hero from "@/components/Hero/Hero";
-
+import Hero from "@/components/Hero";
+import Layout from "@/components/Layout";
 export default function Home() {
   return (
     <>
@@ -11,9 +11,11 @@ export default function Home() {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main className={styles.main}>
-        <Hero />
-      </main>
+      <Layout>
+        <main className={styles.main}>
+          <Hero />
+        </main>
+      </Layout>
     </>
   );
 }
