+++ client/src/styles/globals.scss
@@ -7,6 +7,9 @@ body {
   ::selection {
     color: white;
     background: #1e1c1c;
+    // alt colors
+    // color: #272525;
+    // background: white;
     mix-blend-mode: multiply;
   }
 }
